refactor(Promotions): 프로모션 필드 추가 및 패키지 변경
- 프로모션 파일 입력으로 인해, 필드를 추가했습니다.

- 또한, 모호한 패키지 위치를 domain 바로 하위로 옮겼습니다.
